Make 'update' warn about merge conflicts
git-svn-id: https://shellinabox.googlecode.com/svn/trunk@147 0da03de8-d603-11dd-86c2-0f8696b7b6f9

+++ libhttp/httpconnection.c
@@ -595,7 +595,8 @@ void httpTransfer(struct HttpConnection *http, char *msg, int len) {
           free(msg);
           msg               = compressed;
           len              -= strm.avail_out;
-          snprintf(contentLength, 21, "%20d", len - (ptr - compressed));
+          snprintf(contentLength, 21, "%20lu",
+                   (unsigned long)(len - (ptr - compressed)));
           contentLength[20] = '\r';
         } else {
           free(compressed);
